hw2/task2/cloud-function/requirements.txt
google-cloud-storage
google-cloud-error-reporting